refactored instance params typing pattern
