+++ src/app/(protected)/[group]/[subGroup]/[instance]/(student)/student-overview.tsx
@@ -1,7 +1,5 @@
-export function StudentOverview({
-  params,
-}: {
-  params: { group: string; subGroup: string; instance: string };
-}) {
+import { instanceParams } from "@/lib/validations/params";
+
+export function StudentOverview({ params }: { params: instanceParams }) {
   return <>add student details here</>;
 }
