+++ src/components/stage-button.tsx
@@ -26,7 +26,7 @@ export function StageButton({
         onClick={() => handleSelection(num)}
         className={cn(
           "flex h-16 w-16 items-center justify-center rounded-full bg-accent text-lg font-medium",
-          selectedIdx === num && "border-4 border-amber-500",
+          selectedIdx === num && "border-4 border-sky-500",
           confirmedIdx >= num && "bg-primary text-primary-foreground",
         )}
       >
